Day 8: Part 2
By representing each segment as a prime number we are able to do simple
arithmetic to combine and subtract the different segments.

For example, knowing the values for the products of all the segments
that make up 8 and 9, you can determine the value for the center segment
by dividing the product for 9 by the product for 8.